+++ modules/pytket-pysimplex/_metadata.py
@@ -1,2 +1,2 @@
-__extension_version__ = "0.2.0"
+__extension_version__ = "0.4.0"
 __extension_name__ = "pytket-pysimplex"