+++ modules/pytket-qiskit/tests/backend_test.py
@@ -47,7 +47,7 @@ from pytket.utils.expectations import (
     get_operator_expectation_value,
 )
 from pytket.utils.operators import QubitPauliOperator
-from pytket.utils.results import compare_unitaries
+from pytket.utils.results import compare_statevectors
 from qiskit import IBMQ  # type: ignore
 from qiskit import ClassicalRegister, QuantumCircuit, QuantumRegister
 from qiskit.circuit import Parameter  # type: ignore
@@ -137,7 +137,7 @@ def test_noise() -> None:
         IBMQ.load_account()
 
     provider = IBMQ.providers(hub="ibm-q", group="open")[0]
-    back = provider.get_backend("ibmq_santiago")
+    back = provider.get_backend("ibmq_manila")
 
     noise_model = NoiseModel.from_backend(back)
     n_qbs = 5
@@ -184,7 +184,7 @@ def test_process_characterisation() -> None:
         IBMQ.load_account()
 
     provider = IBMQ.providers(hub="ibm-q", group="open")[0]
-    back = provider.get_backend("ibmq_santiago")
+    back = provider.get_backend("ibmq_manila")
 
     char = process_characterisation(back)
     arch: Architecture = char.get("Architecture", Architecture([]))
@@ -378,8 +378,8 @@ def test_cancellation_ibmq(lima_backend: IBMQBackend) -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_machine_debug(santiago_backend: IBMQBackend) -> None:
-    backend = santiago_backend
+def test_machine_debug(manila_backend: IBMQBackend) -> None:
+    backend = manila_backend
     backend._MACHINE_DEBUG = True
     try:
         c = Circuit(2, 2).H(0).CX(0, 1).measure_all()
@@ -411,8 +411,8 @@ def test_machine_debug(santiago_backend: IBMQBackend) -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_nshots_batching(santiago_backend: IBMQBackend) -> None:
-    backend = santiago_backend
+def test_nshots_batching(manila_backend: IBMQBackend) -> None:
+    backend = manila_backend
     backend._MACHINE_DEBUG = True
     try:
         c1 = Circuit(2, 2).H(0).CX(0, 1).measure_all()
@@ -443,7 +443,7 @@ def test_nshots() -> None:
     if not skip_remote_tests:
         backends.append(
             IBMQEmulatorBackend(
-                "ibmq_santiago", hub="ibm-q", group="open", project="main"
+                "ibmq_manila", hub="ibm-q", group="open", project="main"
             )
         )
     for b in backends:
@@ -479,8 +479,8 @@ def test_pauli_sim() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_default_pass(santiago_backend: IBMQBackend) -> None:
-    b = santiago_backend
+def test_default_pass(manila_backend: IBMQBackend) -> None:
+    b = manila_backend
     for ol in range(3):
         comp_pass = b.default_compilation_pass(ol)
         c = Circuit(3, 3)
@@ -500,7 +500,7 @@ def test_aer_default_pass() -> None:
         IBMQ.load_account()
 
     provider = IBMQ.providers(hub="ibm-q", group="open")[0]
-    back = provider.get_backend("ibmq_santiago")
+    back = provider.get_backend("ibmq_manila")
 
     noise_model = NoiseModel.from_backend(back)
     for nm in [None, noise_model]:
@@ -787,7 +787,7 @@ def test_aer_placed_expectation() -> None:
         IBMQ.load_account()
 
     provider = IBMQ.providers(hub="ibm-q", group="open")[0]
-    back = provider.get_backend("ibmq_santiago")
+    back = provider.get_backend("ibmq_manila")
 
     noise_model = NoiseModel.from_backend(back)
 
@@ -806,7 +806,7 @@ def test_aer_placed_expectation() -> None:
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 def test_ibmq_emulator() -> None:
     b_emu = IBMQEmulatorBackend(
-        "ibmq_santiago", hub="ibm-q", group="open", project="main"
+        "ibmq_manila", hub="ibm-q", group="open", project="main"
     )
     assert b_emu._noise_model is not None
     b_ibm = b_emu._ibmq
@@ -916,13 +916,13 @@ def test_remote_simulator() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_ibmq_mid_measure(santiago_backend: IBMQBackend) -> None:
+def test_ibmq_mid_measure(manila_backend: IBMQBackend) -> None:
     c = Circuit(3, 3).H(1).CX(1, 2).Measure(0, 0).Measure(1, 1)
     c.add_barrier([0, 1, 2])
 
     c.CX(1, 0).H(0).Measure(2, 2)
 
-    b = santiago_backend
+    b = manila_backend
     ps = b.default_compilation_pass(0)
     ps.apply(c)
     # c = b.get_compiled_circuit(c)
@@ -931,9 +931,9 @@ def test_ibmq_mid_measure(santiago_backend: IBMQBackend) -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_compile_x(santiago_backend: IBMQBackend) -> None:
+def test_compile_x(manila_backend: IBMQBackend) -> None:
     # TKET-1028
-    b = santiago_backend
+    b = manila_backend
     c = Circuit(1).X(0)
     for ol in range(3):
         c1 = c.copy()
@@ -960,7 +960,7 @@ def lift_perm(p: Dict[int, int]) -> np.ndarray:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_compilation_correctness(santiago_backend: IBMQBackend) -> None:
+def test_compilation_correctness(manila_backend: IBMQBackend) -> None:
     c = Circuit(5)
     c.H(0).H(1).H(2)
     c.CX(0, 1).CX(1, 2)
@@ -975,7 +975,7 @@ def test_compilation_correctness(santiago_backend: IBMQBackend) -> None:
     c.CX(0, 3).CX(0, 4)
     u_backend = AerUnitaryBackend()
     u = u_backend.run_circuit(c).get_unitary()
-    ibm_backend = santiago_backend
+    ibm_backend = manila_backend
     for ol in range(3):
         p = ibm_backend.default_compilation_pass(optimisation_level=ol)
         cu = CompilationUnit(c)
@@ -993,7 +993,11 @@ def test_compilation_correctness(santiago_backend: IBMQBackend) -> None:
         m_ini = lift_perm(ini)
         m_inv_fin = lift_perm(inv_fin)
 
-        assert compare_unitaries(u, m_inv_fin @ compiled_u @ m_ini)
+        # Note that we do not expect unitary equivalence, because the pass includes
+        # SimplifyInitial which may remove initial gates that do not affect the final
+        # result. However the first columns of the unitaries (i.e. the final
+        # statevectors arising from an initial all-zero state) should match.
+        assert compare_statevectors(u[:, 0], (m_inv_fin @ compiled_u @ m_ini)[:, 0])
 
 
 # pytket-extensions issue #69
@@ -1058,7 +1062,7 @@ def test_postprocess(lima_backend: IBMQBackend) -> None:
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 def test_postprocess_emu() -> None:
-    b = IBMQEmulatorBackend("ibmq_santiago", hub="ibm-q", group="open", project="main")
+    b = IBMQEmulatorBackend("ibmq_manila", hub="ibm-q", group="open", project="main")
     assert b.supports_contextual_optimisation
     c = Circuit(2, 2)
     c.SX(0).SX(1).CX(0, 1).measure_all()
@@ -1108,7 +1112,7 @@ def test_available_devices() -> None:
 def test_backendinfo_serialization1() -> None:
     # https://github.com/CQCL/tket/issues/192
     backend = IBMQEmulatorBackend(
-        "ibmq_santiago", hub="ibm-q", group="open", project="main"
+        "ibmq_manila", hub="ibm-q", group="open", project="main"
     )
     backend_info_json = backend.backend_info.to_dict()
     s = json.dumps(backend_info_json)