+++ modules/pytket-qsharp/_metadata.py
@@ -1,2 +1,2 @@
-__extension_version__ = "0.25.0"
+__extension_version__ = "0.27.0"
 __extension_name__ = "pytket-qsharp"