+++ modules/pytket-quantinuum/tests/api1_test.py
@@ -17,22 +17,27 @@
 # phase.
 
 from io import StringIO
-from typing import Any, Optional, Tuple
+from typing import Any, Dict, List, Optional, Tuple
+from http import HTTPStatus
+from unittest.mock import patch, MagicMock
 import pytest
 
+import requests
 from requests_mock.mocker import Mocker
 
 from pytket.extensions.quantinuum.backends.api_wrappers import QuantinuumAPI
 from pytket.extensions.quantinuum.backends import QuantinuumBackend
 from pytket.circuit import Circuit  # type: ignore
-
+from pytket.architecture import FullyConnected  # type: ignore
 from pytket.extensions.quantinuum.backends.quantinuum import DEFAULT_API_HANDLER
+from pytket.extensions.quantinuum._metadata import __extension_version__
 
 
 def test_default_login_flow(
     requests_mock: Mocker,
     mock_credentials: Tuple[str, str],
     mock_token: str,
+    mock_machine_info: Dict[str, Any],
     monkeypatch: Any,
 ) -> None:
     """Test that when an api_handler is not provided to
@@ -45,7 +50,7 @@ def test_default_login_flow(
 
     DEFAULT_API_HANDLER.delete_authentication()
 
-    fake_device = "H9-27"
+    fake_device = mock_machine_info["name"]
     fake_job_id = "abc-123"
 
     login_route = requests_mock.register_uri(
@@ -69,6 +74,12 @@ def test_default_login_flow(
         "GET",
         f"https://qapi.quantinuum.com/v1/job/{fake_job_id}?websocket=true",
         json={"job": fake_job_id},
+        headers={"Content-Type": "application/json"},
+    )
+    requests_mock.register_uri(
+        "GET",
+        f"https://qapi.quantinuum.com/v1/machine/?config=true",
+        json=[mock_machine_info],
         headers={"Content-Type": "application/json"},
     )
 
@@ -108,6 +119,7 @@ def test_default_login_flow(
 def test_custom_login_flow(
     requests_mock: Mocker,
     mock_token: str,
+    mock_machine_info: Dict[str, Any],
 ) -> None:
     """Test that when an api_handler is provided to
     QuantinuumBackend we use that handler and acquire
@@ -116,7 +128,7 @@ def test_custom_login_flow(
 
     DEFAULT_API_HANDLER.delete_authentication()
 
-    fake_device = "H9-27"
+    fake_device = mock_machine_info["name"]
     fake_job_id = "abc-123"
 
     login_route = requests_mock.register_uri(
@@ -140,6 +152,12 @@ def test_custom_login_flow(
         "GET",
         f"https://qapi.quantinuum.com/v1/job/{fake_job_id}?websocket=true",
         json={"job": fake_job_id},
+        headers={"Content-Type": "application/json"},
+    )
+    requests_mock.register_uri(
+        "GET",
+        f"https://qapi.quantinuum.com/v1/machine/?config=true",
+        json=[mock_machine_info],
         headers={"Content-Type": "application/json"},
     )
 
@@ -179,13 +197,127 @@ def test_custom_login_flow(
     assert job_status_route.call_count == 2  # type: ignore
 
 
+def test_mfa_login_flow(
+    requests_mock: Mocker,
+    mock_credentials: Tuple[str, str],
+    mock_token: str,
+    mock_mfa_code: str,
+    mock_machine_info: Dict[str, Any],
+    monkeypatch: Any,
+) -> None:
+    """Test that the MFA authentication works as expected"""
+
+    DEFAULT_API_HANDLER.delete_authentication()
+
+    fake_device = mock_machine_info["name"]
+
+    def match_mfa_request(request: requests.PreparedRequest) -> bool:
+        return "code" in request.body  # type: ignore
+
+    def match_normal_request(request: requests.PreparedRequest) -> bool:
+        return "code" not in request.body  # type: ignore
+
+    mfa_login_route = requests_mock.register_uri(
+        "POST",
+        "https://qapi.quantinuum.com/v1/login",
+        json={
+            "id-token": mock_token,
+            "refresh-token": mock_token,
+        },
+        headers={"Content-Type": "application/json"},
+        additional_matcher=match_mfa_request,  # type: ignore
+    )
+    normal_login_route = requests_mock.register_uri(
+        "POST",
+        "https://qapi.quantinuum.com/v1/login",
+        json={
+            "error": {"code": 73},
+        },
+        headers={"Content-Type": "application/json"},
+        additional_matcher=match_normal_request,  # type: ignore
+        status_code=HTTPStatus.UNAUTHORIZED,
+    )
+
+    username, pwd = mock_credentials
+    # fake user input from stdin
+    inputs = iter([username + "\n", mock_mfa_code + "\n"])
+    monkeypatch.setattr("builtins.input", lambda msg: next(inputs))
+    monkeypatch.setattr("getpass.getpass", lambda prompt: pwd)
+
+    backend = QuantinuumBackend(
+        device_name=fake_device,
+    )
+    backend.login()
+
+    assert normal_login_route.called_once  # type: ignore
+    # Check that the mfa login has been invoked
+    assert mfa_login_route.called_once  # type: ignore
+    assert backend._api_handler._cred_store.id_token is not None
+    assert backend._api_handler._cred_store.refresh_token is not None
+
+
+@patch("pytket.extensions.quantinuum.backends.api_wrappers.microsoft_login")
+def test_federated_login(
+    mock_microsoft_login: MagicMock,
+    requests_mock: Mocker,
+    mock_credentials: Tuple[str, str],
+    mock_token: str,
+    mock_ms_provider_token: str,
+    mock_machine_info: Dict[str, Any],
+) -> None:
+    """Test that the federated authentication works as expected"""
+    DEFAULT_API_HANDLER.delete_authentication()
+
+    fake_device = mock_machine_info["name"]
+
+    backend = QuantinuumBackend(
+        device_name=fake_device,
+        provider="microsoft",
+    )
+    login_route = requests_mock.register_uri(
+        "POST",
+        "https://qapi.quantinuum.com/v1/login",
+        json={
+            "id-token": mock_token,
+            "refresh-token": mock_token,
+        },
+        headers={"Content-Type": "application/json"},
+    )
+    mock_microsoft_login.return_value = (mock_credentials[0], mock_ms_provider_token)
+    backend.login()
+
+    mock_microsoft_login.assert_called_once()
+    assert login_route.called_once  # type: ignore
+    assert backend._api_handler._cred_store.id_token is not None
+    assert backend._api_handler._cred_store.refresh_token is not None
+
+
+def test_federated_login_wrong_provider(
+    mock_machine_info: Dict[str, Any],
+) -> None:
+    """Test that the federated authentication works as expected"""
+    DEFAULT_API_HANDLER.delete_authentication()
+
+    fake_device = mock_machine_info["name"]
+
+    backend = QuantinuumBackend(
+        device_name=fake_device,
+        provider="wrong provider",
+    )
+    with pytest.raises(RuntimeError) as e:
+        backend.login()
+        err_msg = "Unsupported provider for login"
+        assert err_msg in str(e.value)
+
+
 @pytest.mark.parametrize(
     "chosen_device,max_batch_cost",
-    [("H1", 300), ("H1", None), ("H1-1", 300), ("H1-1", None)],
+    [("H1", 300), ("H1-1", 300), ("H1-1", None)],
 )
 def test_device_family(
     requests_mock: Mocker,
     mock_quum_api_handler: QuantinuumAPI,
+    sample_machine_infos: List[Dict[str, Any]],
     chosen_device: str,
     max_batch_cost: Optional[int],
 ) -> None:
@@ -209,6 +341,13 @@ def test_device_family(
         headers={"Content-Type": "application/json"},
     )
 
+    requests_mock.register_uri(
+        "GET",
+        f"https://qapi.quantinuum.com/v1/machine/?config=true",
+        json=sample_machine_infos,
+        headers={"Content-Type": "application/json"},
+    )
+
     family_backend = QuantinuumBackend(
         device_name=chosen_device,
     )
@@ -228,7 +367,7 @@ def test_device_family(
     if requests_mock.last_request:
         submitted_json = requests_mock.last_request.json()
 
-    if chosen_device == "H1" and max_batch_cost is None:
+    if chosen_device == "H1":
         assert "batch-exec" not in submitted_json.keys()
         assert "batch-end" not in submitted_json.keys()
     else:
@@ -239,6 +378,7 @@ def test_device_family(
 def test_resumed_batching(
     requests_mock: Mocker,
     mock_quum_api_handler: QuantinuumAPI,
+    sample_machine_infos: Dict[str, Any],
 ) -> None:
     """Test that you can resume using a batch."""
 
@@ -255,6 +395,12 @@ def test_resumed_batching(
         "GET",
         f"https://qapi.quantinuum.com/v1/job/{fake_job_id}?websocket=true",
         json={"job": fake_job_id},
+        headers={"Content-Type": "application/json"},
+    )
+    requests_mock.register_uri(
+        "GET",
+        f"https://qapi.quantinuum.com/v1/machine/?config=true",
+        json=sample_machine_infos,
         headers={"Content-Type": "application/json"},
     )
 
@@ -287,3 +433,39 @@ def test_resumed_batching(
         submitted_json = requests_mock.last_request.json()
     assert submitted_json["batch-exec"] == backend.get_jobid(h1)
     assert "batch-end" in submitted_json
+
+
+def test_available_devices(
+    requests_mock: Mocker,
+    mock_quum_api_handler: QuantinuumAPI,
+    mock_machine_info: Dict[str, Any],
+) -> None:
+
+    requests_mock.register_uri(
+        "GET",
+        f"https://qapi.quantinuum.com/v1/machine/?config=true",
+        json=[mock_machine_info],
+        headers={"Content-Type": "application/json"},
+    )
+
+    devices = QuantinuumBackend.available_devices(_api_handler=mock_quum_api_handler)
+    assert len(devices) == 1
+    backinfo = devices[0]
+
+    assert backinfo.device_name == mock_machine_info["name"]
+    assert backinfo.architecture == FullyConnected(mock_machine_info["n_qubits"])
+    assert backinfo.version == __extension_version__
+    assert backinfo.supports_fast_feedforward == True
+    assert backinfo.supports_midcircuit_measurement == True
+    assert backinfo.supports_reset == True
+    assert backinfo.misc == {
+        "n_classical_registers": 50,
+        "n_shots": 10000,
+        "system_family": "mock_family",
+        "system_type": "hardware",
+        "emulator": "H9-27E",
+        "syntax_checker": "H9-27SC",
+        "batching": True,
+        "wasm": True,
+    }
+    assert backinfo.name == "QuantinuumBackend"