+++ modules/pytket-cirq/docs/changelog.rst
@@ -1,10 +1,19 @@
 Changelog
 ~~~~~~~~~
 
-0.24.0 (Unreleased)
--------------------
+0.25.0 (July 2022)
+------------------
+
+* Updated pytket version requirement to 1.4.
+* Updated cirq version requirement to 0.15.
+* Changed `process_characterisation()` so that it expects a `GridDevice` instead
+  of a `SerializableDevice`.
+
+0.24.0 (June 2022)
+------------------
 
 * `cirq_to_tk()` and `tk_to_cirq()` now properly handle circuits with `NamedQubit`.
+* Updated pytket version requirement to 1.3.
 
 0.23.0 (May 2022)
 -----------------