+++ modules/pytket-ionq/docs/changelog.rst
@@ -1,6 +1,16 @@
 Changelog
 ~~~~~~~~~
 
+0.20.0 (July 2022)
+------------------
+
+* Updated pytket version requirement to 1.4.
+
+0.19.0 (June 2022)
+------------------
+
+* Updated pytket version requirement to 1.3.
+
 0.18.0 (May 2022)
 -----------------