+++ modules/pytket-qsharp/docs/changelog.rst
@@ -1,6 +1,17 @@
 Changelog
 ~~~~~~~~~
 
+0.27.0 (July 2022)
+------------------
+
+* Updated pytket version requirement to 1.4.
+* Updated qsharp version requirement to 0.25.
+
+0.26.0 (June 2022)
+------------------
+
+* Updated pytket version requirement to 1.3.
+
 0.25.0 (May 2022)
 -----------------